+++ src/accessories/accessory.setup.js
@@ -71,9 +71,9 @@ const refreshCache = async (device, bravia, storagePath) => {
         }
       });
 
-      /*tvCache.commands = tvCache.commands.filter(
+      tvCache.commands = tvCache.commands.filter(
         (cachedCommand) => cachedCommand && commands.find((command) => command && command.value === cachedCommand.value)
-      );*/
+      );
 
       //REFRESH INPUTS
       inputs.forEach((exInput) => {
@@ -168,48 +168,54 @@ const Setup = async (devices, tvConfigs, storagePath) => {
 
             //REFRESH APPS
             apps.forEach((app) => {
-              const cachedApp = tvCache.apps.find((cachedApp) => cachedApp.name === app.name);
+              const cachedApp = tvCache.apps.find((cachedApp) => cachedApp && cachedApp.name === app.name);
               if (!cachedApp) {
                 tvCache.apps.push(app);
               }
             });
 
-            tvCache.apps = tvCache.apps.filter((cachedApp) => apps.find((app) => app.name === cachedApp.name));
+            tvCache.apps = tvCache.apps.filter((cachedApp) => apps.find((app) => app && app.name === cachedApp.name));
 
             //REFRESH CHANNELS
             channels.forEach((channel) => {
-              const cachedChannel = tvCache.channels.find((cachedChannel) => cachedChannel.uri === channel.uri);
+              const cachedChannel = tvCache.channels.find(
+                (cachedChannel) => cachedChannel && cachedChannel.uri === channel.uri
+              );
               if (!cachedChannel) {
                 tvCache.channels.push(channel);
               }
             });
 
-            tvCache.channels = tvCache.channels.filter((cachedChannel) =>
-              channels.find((channel) => channel.uri === cachedChannel.uri)
+            tvCache.channels = tvCache.channels.filter(
+              (cachedChannel) => cachedChannel && channels.find((channel) => channel.uri === cachedChannel.uri)
             );
 
             //REFRESH COMMANDS
             commands.forEach((command) => {
-              const cachedCommand = tvCache.commands.find((cachedCommand) => cachedCommand.value === command.value);
+              const cachedCommand = tvCache.commands.find(
+                (cachedCommand) => cachedCommand && cachedCommand.value === command.value
+              );
               if (!cachedCommand) {
                 tvCache.commands.push(command);
               }
             });
 
-            /*tvCache.commands = tvCache.commands.filter((cachedCommand) =>
-              commands.find((command) => command.value === cachedCommand.value)
-            );*/
+            tvCache.commands = tvCache.commands.filter(
+              (cachedCommand) => cachedCommand && commands.find((command) => command.value === cachedCommand.value)
+            );
 
             //REFRESH INPUTS
             inputs.forEach((exInput) => {
-              const cachedExInput = tvCache.inputs.find((cachedExInput) => cachedExInput.uri === exInput.uri);
+              const cachedExInput = tvCache.inputs.find(
+                (cachedExInput) => cachedExInput && cachedExInput.uri === exInput.uri
+              );
               if (!cachedExInput) {
                 tvCache.inputs.push(exInput);
               }
             });
 
             /*tvCache.inputs = tvCache.inputs.filter((cachedExInput) =>
-              inputs.find((exInput) => exInput.uri === cachedExInput.uri)
+              inputs.find((exInput) => exInput && exInput.uri === cachedExInput.uri)
             );*/
           } else {
             //NEW